rfac: change level convention, rename files
